requirements.txt
discord
python-dotenv
beautifulsoup4
requests
tabulate